Implemented local storage
I have written the local storage integration myself instead of using the third party library previously used. This removes an unnecessary 6KB library and much redundant functionality which we aren't interested in. I have mimicked the structure of the objects in the third party library so that we can use it again in the future if the need arises.